+++ Backend/App/__init__.py
@@ -5,8 +5,6 @@ from datetime import timedelta
 from flask import Flask
 from flask_cors import CORS
 from flask_jwt_extended import JWTManager
-from flask_limiter import Limiter
-from flask_limiter.util import get_remote_address
 from flask_socketio import SocketIO
 
 from App.extensions import limiter