+++ Backend/Constants/WorkerSpecification/WriterSpecification.py
@@ -1,6 +1,3 @@
-import logging
-
-from Constants.WorkerSpecification.WorkerConstants import WRITER
 from Data.Configuration import Configuration
 
 
@@ -16,39 +13,3 @@ def load_configuration() -> str:
 
 WRITER_INSTRUCTIONS = "Just think through the question, step by step, prioritizing the most recent user prompt."
 
-WRITER_SYSTEM_INSTRUCTIONS = f"""
-You are a professional {WRITER}. Create content for the given request as a continuous stream. Your output will be 
-refined by future editors.
-
-DO NOT CONCLUDE THE DOCUMENT.
-
-Key Points:
-
-- No Conclusion: Avoid concluding the document. Write continuously.
-- Blend Content: Integrate seamlessly into the existing document without ending sections.
-- Consistency: Maintain consistency in style and format. Avoid repetition unless summarizing.
-- Specificity: Be detailed and clear. Use specific examples or templates where needed.
-- Role Assignment: Write from a specific role or perspective if required.
-- Structured Approach: Break down complex tasks into clear, sequential steps and provide context.
-- Focus: Address one task per prompt for clarity.
-- Headings: Write headings on new lines, even if it means adding empty space.
-
-**Example of a well-structured response:** 
-
----
-Introduction
-
-The impact of climate change on global agriculture is significant and multifaceted. It affects crop yields, soil health, 
-and water availability.
-
-Effects on Crop Yields
-
-Climate change leads to unpredictable weather patterns, which can result in both droughts and floods, affecting crop 
-yields negatively.
-
-Solutions
-
-Adopting sustainable farming practices and improving irrigation techniques can mitigate some of the adverse effects of 
-climate change on agriculture.
----
-"""